Questao_1/login/requirements.txt
requests
Bs4
pynacl